+++ main/resources/mcmod.info
@@ -1,14 +1,14 @@
 [
 {
-  "modid": "examplemod",
-  "name": "Example Mod",
-  "description": "Example placeholder mod.",
-  "version": "${version}",
-  "mcversion": "${mcversion}",
+  "modid": "vme",
+  "name": "Virgin Modding Environment",
+  "description": "Where new things are learnt and coding cherries popped.",
+  "version": "1.0",
+  "mcversion": "1.12.1",
   "url": "",
   "updateUrl": "",
-  "authorList": ["ExampleDude"],
-  "credits": "The Forge and FML guys, for making this example",
+  "authorList": ["LunariusH","Gogtabi"],
+  "credits": "Harry, CMJ, LunariusH",
   "logoFile": "",
   "screenshots": [],
   "dependencies": []
